기본적인 회초리 코드 수정 - apiResponse 미적용 - test 미적용 - ExceptionCode 미적용
