+++ src/main/java/younah/TodoRefactor/domain/todo/controller/BusinessTodoController.java
@@ -5,15 +5,21 @@ import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import younah.TodoRefactor.domain.todo.service.WithdrawTodoService;
+import younah.TodoRefactor.domain.todo.service.BusinessTodoService;
+
 
 @RestController
-@RequestMapping("/todos")
+@RequestMapping("/todos/{todoId}")
 @RequiredArgsConstructor
-class WithdrawTodoController {
-    private final WithdrawTodoService service;
+public class BusinessTodoController {
+    private final BusinessTodoService service;
 
-    @PatchMapping("/withdraw/{todoId}")
+    @PatchMapping("/complete")
+    void complete(@PathVariable long todoId){
+        service.complete(todoId);
+    }
+
+    @PatchMapping("/withdraw")
     void withdraw(@PathVariable long todoId){
         service.withdraw(todoId);
     }
