+++ src/main/java/younah/TodoRefactor/domain/todo/controller/DeleteTodoController.java
@@ -5,25 +5,16 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import younah.TodoRefactor.domain.todo.dto.TodoDto;
 import younah.TodoRefactor.domain.todo.service.DeleteTodoService;
 
-import java.time.LocalDateTime;
-
 @RestController
-@RequestMapping("/todos")
+@RequestMapping("/todos/{todoId}")
 @RequiredArgsConstructor
 public class DeleteTodoController {
     private final DeleteTodoService service;
 
-    @DeleteMapping("/{todoId}")
-    Response delete(@PathVariable long todoId){
-        TodoDto todoDto = service.remove(todoId);
-        return new Response(todoDto.id(), todoDto.deletedAt());
+    @DeleteMapping
+    void delete(@PathVariable long todoId){
+        service.remove(todoId);
     }
-
-    record Response(
-            Long id,
-            LocalDateTime deletedAt
-    ){}
 }
