+++ src/main/java/younah/TodoRefactor/domain/todo/controller/UpdateTodoController.java
@@ -4,35 +4,24 @@ import jakarta.validation.constraints.NotBlank;
 import lombok.RequiredArgsConstructor;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-
-import younah.TodoRefactor.domain.todo.dto.TodoDto;
-import younah.TodoRefactor.domain.todo.entity.Todo;
 import younah.TodoRefactor.domain.todo.service.UpdateTodoService;
 
 //request가 존재한다면, 해당 request 안에 requirement로 변환해주는 메소드를 지니고 있어야 함
 //동시에, 해당 service에 requirement가 선언되어 있어야 함 (record)
 
 @RestController
-@RequestMapping("/todos")
+@RequestMapping("/todos/{todoId}")
 @RequiredArgsConstructor
 @Validated
 class UpdateTodoController {
 
     private final UpdateTodoService service;
 
-    @PatchMapping("/{todoId}")
-    Response updateTodo(@RequestBody Request request,
+    @PutMapping
+    void updateTodo(@RequestBody Request request,
                         @PathVariable long todoId){
-
         //requirement 변환
-        TodoDto responseDto = service.update(todoId, request.toRequirement());
-
-        //response 변환
-        return new Response(
-                responseDto.id(),
-                responseDto.content(),
-                responseDto.status()
-        );
+        service.update(todoId, request.toRequirement());
     }
 
     public record Request(
@@ -43,12 +32,4 @@ class UpdateTodoController {
             return new UpdateTodoService.Requirement(content);
         }
     }
-
-    public record Response(
-            Long id,
-            String content,
-            Todo.TodoStatus status
-    ){}
-
-
 }
