+++ src/main/java/younah/TodoRefactor/domain/todo/service/DeleteTodoService.java
@@ -3,7 +3,6 @@ package younah.TodoRefactor.domain.todo.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import younah.TodoRefactor.domain.todo.dto.TodoDto;
 import younah.TodoRefactor.domain.todo.entity.Todo;
 import younah.TodoRefactor.domain.todo.repository.TodoRepository;
 import younah.TodoRefactor.global.exception.BusinessLogicException;
@@ -15,13 +14,11 @@ public class DeleteTodoService {
     private final TodoRepository todoRepo;
 
     @Transactional
-    public TodoDto remove(long todoId){
+    public void remove(long todoId){
         Todo targetTodo = todoRepo.findById(todoId)
                 .orElseThrow(() -> new BusinessLogicException(ExceptionCode.TODO_NOT_EXSIST));
 
         targetTodo.remove();
         todoRepo.save(targetTodo);
-
-        return TodoDto.fromEntity(targetTodo);
     }
 }
