+++ src/main/java/younah/TodoRefactor/domain/todo/service/UpdateTodoService.java
@@ -3,7 +3,6 @@ package younah.TodoRefactor.domain.todo.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import younah.TodoRefactor.domain.todo.dto.TodoDto;
 import younah.TodoRefactor.domain.todo.entity.Todo;
 import younah.TodoRefactor.domain.todo.repository.TodoRepository;
 import younah.TodoRefactor.global.exception.BusinessLogicException;
@@ -15,27 +14,19 @@ public class UpdateTodoService {
     private final TodoRepository todoRepo;
 
     @Transactional
-    public TodoDto update(long todoId, Requirement requirement){
-
-        Todo newTodo = requirement.toEntity();
+    public void update(long todoId, Requirement requirement){
 
         Todo targetTodo = todoRepo.findById(todoId)
                 .orElseThrow(() -> new BusinessLogicException(ExceptionCode.TODO_NOT_EXSIST));
 
-        targetTodo.update(newTodo.getContent());
-        //수정을 위해 받아온 requirement를 entity로 변환해야 하는가? ??
-        // ㄴ> 음 ... 값만 받는데 그걸 requestBody로 받기 위해서 이런거임 ... 얘를 entity로 바꿀 필요는 없다 생각함
-            //ㄴ> 이래놓고 다시 수정 컨트롤러 서비스 침범 멈춰
-        todoRepo.save(targetTodo);
+        targetTodo.update(requirement.content());
+        //수정을 위해 받아온 requirement를 entity로 변환해야 하는가? ?? ㄴㄴ 바로 content 뽑아쓰기
+        //어차피 service 객체로 오기 위한 requirement로 변환했음 !!
 
-        return TodoDto.fromEntity(targetTodo);
+        todoRepo.save(targetTodo);
     }
 
     public record Requirement(
             String content
-    ){
-        public Todo toEntity(){
-        return new Todo(content);
-    }
-    }
+    ){ }
 }
