--- src/main/java/younah/TodoRefactor/domain/todo/service/WithdrawTodoService.java
@@ -1,24 +1,0 @@
-package younah.TodoRefactor.domain.todo.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import younah.TodoRefactor.domain.todo.entity.Todo;
-import younah.TodoRefactor.domain.todo.repository.TodoRepository;
-import younah.TodoRefactor.global.exception.BusinessLogicException;
-import younah.TodoRefactor.global.exception.ExceptionCode;
-
-@Service
-@RequiredArgsConstructor
-public class WithdrawTodoService {
-    private final TodoRepository todoRepo;
-
-    @Transactional
-    public void withdraw(long todoId){
-        Todo targetTodo = todoRepo.findById(todoId)
-                .orElseThrow(() -> new BusinessLogicException(ExceptionCode.TODO_NOT_EXSIST));
-
-        targetTodo.withdraw();
-        todoRepo.save(targetTodo);
-    }
-}
